% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-energy-requirements.R
\name{ADG_from_empty_gain_weight}
\alias{ADG_from_empty_gain_weight}
\title{Average daily gain calculated from the empty weight gain.}
\usage{
ADG_from_empty_gain_weight(EGW = 0.45, liquid_diet_only = TRUE, weaned = FALSE)
}
\arguments{
\item{EGW}{Empty weight gain (kg).}

\item{liquid_diet_only}{Argument to define whether the calves have only milk available or also forage.}

\item{weaned}{Argument to define whether the calves are weaned or not.}
}
\value{
The ADG from empty weight gain.
}
\description{
Average daily gain calculated from the empty weight gain.
}
