% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{efMPgain}
\alias{efMPgain}
\title{NASEM (2021) equation to predict the efficiency of converting metabolizable protein for gain (MPg) to net protein (NPg).}
\usage{
efMPgain(BW = 50, mature_weight = 700)
}
\arguments{
\item{BW}{Body weight (kg).}

\item{mature_weight}{Mature weight (kg).}
}
\value{
The efficiency of converting MPg to NPg.
}
\description{
NASEM (2021) equation to predict the efficiency of converting metabolizable protein for gain (MPg) to net protein (NPg).
}
\examples{
efMPgain(0.16, 0.29)
}
