% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-energy-requirements.R
\name{empty_body_weight_gain_from_re}
\alias{empty_body_weight_gain_from_re}
\title{NASEM (2021) equation tom predict the empty body weight gain based on the retained energy available.}
\usage{
empty_body_weight_gain_from_re(retained_energy = 0.96, EBW = 42.3)
}
\arguments{
\item{retained_energy}{Retained energy (Mcal).}

\item{EBW}{Empty body weight (kg)}
}
\value{
The empty body weight gain based on the RE.
}
\description{
NASEM (2021) equation tom predict the empty body weight gain based on the retained energy available.
}
\examples{
empty_body_weight_gain_from_re(retained_energy = 1, EBW = 45)
}
