% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-energy-requirements.R
\name{empty_weight_gain}
\alias{empty_weight_gain}
\title{NASEM (2021) equation to predict the calf empty body gain.}
\usage{
empty_weight_gain(ADG = 0.45, liquid_diet_only = TRUE, weaned = FALSE)
}
\arguments{
\item{ADG}{Average daily gain (kg).}

\item{liquid_diet_only}{Argument to define whether the calves have only milk available or also forage.}

\item{weaned}{Argument to define whether the calves are weaned or not.}
}
\value{
The empty body gain (kg).
}
\description{
NASEM (2021) equation to predict the calf empty body gain.
}
\examples{
empty_weight_gain(ADG = 0.450, liquid_diet_only = TRUE, weaned = FALSE)
}
