% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{endogenous_urinary_cp}
\alias{endogenous_urinary_cp}
\title{NASEM (2021) equation to predict the endogenous urinary crude protein (CP) losses - Eq. 10-11.}
\usage{
endogenous_urinary_cp(BW = 40)
}
\arguments{
\item{BW}{Body weight (kg).}
}
\value{
Endogenous urinary CP losses (g/d).
}
\description{
NASEM (2021) equation to predict the endogenous urinary crude protein (CP) losses - Eq. 10-11.
}
\examples{
endogenous_urinary_cp(BW = 45)
}
