% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-utils.R
\name{filtering_strings}
\alias{filtering_strings}
\title{Filtering the strings in a pdf file - functiong used in the starter composition module.}
\usage{
filtering_strings(vector, text_pattern)
}
\arguments{
\item{vector}{Vector containing the pdf file content.}

\item{text_pattern}{Text pattern to be searched in the pdf file.}
}
\value{
The filtered strings.
}
\description{
Filtering the strings in a pdf file - functiong used in the starter composition module.
}
