% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-body-composition.R
\name{gain_composition}
\alias{gain_composition}
\title{NASEM (2021) equation to predict the gain composition of dairy calves.}
\usage{
gain_composition(retained_energy = 1.456)
}
\arguments{
\item{retained_energy}{Retained energy (Mcal/d)}
}
\value{
A list with the predicted energy gained in the form of fat and protein in the gain.
}
\description{
NASEM (2021) equation to predict the gain composition of dairy calves.
}
\examples{
gain_composition(retained_energy = 1.45)
}
