% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_ca_req}
\alias{get_ca_req}
\title{Estimates the calf calcium adequate intakes (AIs) - NASEM (2021) Eq. 10-19.}
\usage{
get_ca_req(EBW = 65, EBG = 0.45)
}
\arguments{
\item{EBW}{Empty body weight (kg)}

\item{EBG}{Empty body gain (day)}
}
\value{
Ca adequate intakes (g/day).
}
\description{
Estimates the calf calcium adequate intakes (AIs) - NASEM (2021) Eq. 10-19.
}
\examples{
get_ca_req(EBW = 65, EBG = 0.45)
}
