% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_cu_req}
\alias{get_cu_req}
\title{Estimates the calf sulfur adequate intakes (AI) - NASEM (2021) Eq. 10-24.}
\usage{
get_cu_req(BW = 65, ADG = 0.45)
}
\arguments{
\item{BW}{Body weight (kg).}

\item{ADG}{Average daily gain (kg).}
}
\value{
S adequate intakes (mg/day).
}
\description{
Estimates the calf sulfur adequate intakes (AI) - NASEM (2021) Eq. 10-24.
}
\examples{
get_s_req(BW = 65, ADG = 0.45)
}
