% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_fe_req}
\alias{get_fe_req}
\title{Estimates the calf iron adequate intakes (AI) - NASEM (2021) Eq. 10-25.}
\usage{
get_fe_req(ADG = 0.5)
}
\arguments{
\item{ADG}{Average daily gain (kg).}
}
\value{
Fe adequate intakes (mg/day).
}
\description{
Estimates the calf iron adequate intakes (AI) - NASEM (2021) Eq. 10-25.
}
\examples{
get_fe_req(ADG = 0.5)
}
