% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_k_req}
\alias{get_k_req}
\title{Estimates the calf potassium adequate intakes (AI) - NASEM (2021) Eq. 10-22.}
\usage{
get_k_req(EBW = 65, EBG = 0.45)
}
\arguments{
\item{EBW}{Empty body weight (kg).}

\item{EBG}{Empty body gain (day).}
}
\value{
K adequate intakes (g/day).
}
\description{
Estimates the calf potassium adequate intakes (AI) - NASEM (2021) Eq. 10-22.
}
\examples{
get_k_req(EBW = 65, EBG = 0.45)
}
