% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_mg_req}
\alias{get_mg_req}
\title{Estimates the calf magnesium adequate intakes (AI) - NASEM (2021) Eq. 10-21.}
\usage{
get_mg_req(EBW = 65, EBG = 0.45)
}
\arguments{
\item{EBW}{Empty body weight (kg).}

\item{EBG}{Empty body gain (day).}
}
\value{
Mg adequate intakes (g/day).
}
\description{
Estimates the calf magnesium adequate intakes (AI) - NASEM (2021) Eq. 10-21.
}
\examples{
get_mg_req(EBW = 65, EBG = 0.45)
}
