% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_mn_req}
\alias{get_mn_req}
\title{Estimates the calf manganese adequate intakes (AI) - NASEM (2021) Eq. 10-26.}
\usage{
get_mn_req(BW = 65, ADG = 0.45)
}
\arguments{
\item{BW}{Body weight (kg).}

\item{ADG}{Average daily gain (kg).}
}
\value{
Mn adequate intakes (mg/day).
}
\description{
Estimates the calf manganese adequate intakes (AI) - NASEM (2021) Eq. 10-26.
}
\examples{
get_mn_req(BW = 65, ADG = 0.45)
}
