% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_p_req}
\alias{get_p_req}
\title{Estimates the calf phosphorus adequate intakes (AI) - NASEM (2021) Eq. 10-20.}
\usage{
get_p_req(EBW = 65, EBG = 0.45)
}
\arguments{
\item{EBW}{Empty body weight (kg).}

\item{EBG}{Empty body gain (day).}
}
\value{
P adequate intakes (g/day).
}
\description{
Estimates the calf phosphorus adequate intakes (AI) - NASEM (2021) Eq. 10-20.
}
\examples{
get_p_req(EBW = 65, EBG = 0.45)
}
