% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mineral-requirements.R
\name{get_zn_req}
\alias{get_zn_req}
\title{Estimates the calf zinc adequate intakes (AI) - NASEM (2021) Eq. 10-27.}
\usage{
get_zn_req(DMI = 65, ADG = 0.45)
}
\arguments{
\item{DMI}{Dry matter intake (kg/day).}

\item{ADG}{Average daily gain (kg).}
}
\value{
Zn adequate intakes (mg/day).
}
\description{
Estimates the calf zinc adequate intakes (AI) - NASEM (2021) Eq. 10-27.
}
\examples{
get_zn_req(DMI = 65, ADG = 0.45)
}
