% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-energy-requirements.R
\name{kg_solids}
\alias{kg_solids}
\title{NASEM (2021) equation to predict the efficiency of use of the ME from the starter.}
\usage{
kg_solids(ME = 3)
}
\arguments{
\item{ME}{Metabolizable energy (Mcal).}
}
\value{
Efficiency of use of the ME from starter.
}
\description{
NASEM (2021) equation to predict the efficiency of use of the ME from the starter.
}
\examples{
kg_solids(ME = 3)
}
