% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-energy-requirements.R
\name{km}
\alias{km}
\title{NASEM (2021) equation to predict the efficiency of use of the metabolizable energy.}
\usage{
km(ME = 3.5, liquid_diet_only = TRUE, weaned = FALSE)
}
\arguments{
\item{ME}{Metabolizable energy (Mcal/kg DM).}

\item{liquid_diet_only}{Argument to define whether the calves have only milk available or also forage.}

\item{weaned}{Argument to define whether the calves are weaned or not.}
}
\value{
The efficiency of use of the ME.
}
\description{
NASEM (2021) equation to predict the efficiency of use of the metabolizable energy.
}
\examples{
km(ME = 3.5, liquid_diet_only = TRUE, weaned = FALSE)
}
