% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plots.R
\name{line_plotly}
\alias{line_plotly}
\title{Line plot with two Y axis.}
\usage{
line_plotly(
  data = datasets::iris,
  x_axis = "Sepal.Length",
  y_axis = "Petal.Length",
  y_axis_2 = "Petal.Width",
  x_label = "Sepal length",
  y_label = "Petal Length",
  y_axis_2_label = "Petal.Width",
  legend_title = "Species",
  plot_title = " "
)
}
\arguments{
\item{data}{The data frame to be used.}

\item{x_axis}{The X axis to be plotted.}

\item{y_axis}{The first Y axis to be plotted.}

\item{y_axis_2}{The second (right side) Y axis to be plotted.}

\item{x_label}{The X axis label.}

\item{y_label}{The first Y axis label.}

\item{y_axis_2_label}{The second (right side) label.}

\item{plot_title}{The title of the plot.}
}
\value{
A plot with two Y axis.
}
\description{
Line plot with two Y axis.
}
