% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{metabolic_fecal_cp}
\alias{metabolic_fecal_cp}
\title{NASEM (2021) equation to predict the metabolic fecal crude protein (CP) losses - Eq. 10-12.}
\usage{
metabolic_fecal_cp(liqDietDMI = 0.7, solidDietDMI = 0.4)
}
\arguments{
\item{liqDietDMI}{Liquid feed dry matter intake (DMI; kg/d).}

\item{solidDietDMI}{Solid feed dry matter intake (DMI; kg/d).}
}
\value{
Metabolic fecal CP losses (g/d).
}
\description{
NASEM (2021) equation to predict the metabolic fecal crude protein (CP) losses - Eq. 10-12.
}
\examples{
metabolic_fecal_cp(liqDietDMI   = 0.7, solidDietDMI = 0.4)
}
