% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-milk-allowance-program.R
\name{milk_allowance_program}
\alias{milk_allowance_program}
\title{Allows the user to simulate different nutritional plans for calves.}
\usage{
milk_allowance_program(
  weaning_age = 70,
  nutritional_plans = 4,
  milk_allowance = c(6, 4, 2, 1),
  ages_of_change = c(25, 35, 45)
)
}
\arguments{
\item{weaning_age}{Weaning age (days).}

\item{nutritional_plans}{Number of nutritional plans.}

\item{milk_allowance}{Vector with the milk allowance according to the nutritional plan (kg/day).}

\item{ages_of_change}{Vector with the changes in the milk allowance (days).}
}
\value{
A vector with the milk allowance program.
}
\description{
Allows the user to simulate different nutritional plans for calves.
}
