% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-milk-composition-energetic-value.R
\name{milk_metabolizable_energy}
\alias{milk_metabolizable_energy}
\title{NASEM (2021) equation to predict the milk metabolizable energy - Eq. 10-16.}
\usage{
milk_metabolizable_energy(
  protein = 3.2,
  fat = 3.8,
  ashes = 1,
  total_solids = 12
)
}
\arguments{
\item{protein}{Milk protein content (\%; as is).}

\item{fat}{Milk fat content (\%; as is).}

\item{ashes}{Milk ashes content (\%; as is).}

\item{total_solids}{Milk total solids (\%).}
}
\value{
Milk metabolizable energy (Mcal/l).
}
\description{
NASEM (2021) equation to predict the milk metabolizable energy - Eq. 10-16.
}
\examples{
milk_metabolizable_energy(protein = 3.18, fat = 3.85, ashes = 0.787, total_solids = 12.5)
}
