% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-milk-composition-energetic-value.R
\name{milk_replacer_metabolizable_energy}
\alias{milk_replacer_metabolizable_energy}
\title{NASEM (2021) equation to predict the milk replacer metabolizable energy - Eq. 10-16.}
\usage{
milk_replacer_metabolizable_energy(
  protein = 21.7,
  fat = 18.6,
  ashes = 7.3,
  total_solids = 96.8
)
}
\arguments{
\item{protein}{Milk protein content (\%; DM).}

\item{fat}{Milk fat content (\%; DM).}

\item{ashes}{Milk ashes content (\%; DM).}

\item{total_solids}{Milk total solids (\%; DM).}
}
\value{
Milk replacer metabolizable energy (Mcal/l).
}
\description{
NASEM (2021) equation to predict the milk replacer metabolizable energy - Eq. 10-16.
}
\examples{
milk_replacer_metabolizable_energy(protein = 21.7, fat = 18.6, ashes = 7.3, total_solids = 96.8)
}
