% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{MPmain}
\alias{MPmain}
\title{NASEM (2021) equation to predict the total maintenance net protein (NP) requirement - Eq. 10-13.}
\usage{
MPmain(scurf_cp = 2, endogenous_urinary_cp = 17, metabolic_fecal_cp = 17)
}
\arguments{
\item{scurf_cp}{Total scurf CP losses (g/d).}

\item{endogenous_urinary_cp}{Endogenous urinary CP losses (g/d).}

\item{metabolic_fecal_cp}{Metabolic fecal CP losses (g/d).}
}
\value{
The total maintenance net protein (NP) requirement (g/d)
}
\description{
NASEM (2021) equation to predict the total maintenance net protein (NP) requirement - Eq. 10-13.
}
\examples{
MPmain(scurf_cp = 2, endogenous_urinary_cp = 17, metabolic_fecal_cp  = 17)
}
