% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-energy-requirements.R
\name{NEm}
\alias{NEm}
\title{NASEM (2021) equation to predict the energy requirements for maintenance (Mcal/day).}
\usage{
NEm(EBW = 50, weaned = FALSE)
}
\arguments{
\item{EBW}{Empty body weight (kg).}

\item{weaned}{Argument to define whether the calves are weaned or not.}
}
\value{
The energy requirement for maintenance (Mcal/day).
}
\description{
NASEM (2021) equation to predict the energy requirements for maintenance (Mcal/day).
}
\examples{
NEm(EBW = 50, weaned = FALSE)
}
