% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{net_protein_gain}
\alias{net_protein_gain}
\title{NASEM (2021) equation to predict the net protein requirement for growth - Eq. 10-14.}
\usage{
net_protein_gain(EBW_gain = 0.4, retained_energy = 0.9)
}
\arguments{
\item{EBW_gain}{Empty body weight gain (kg/d).}

\item{retained_energy}{retained energy (Mcal/d).}
}
\value{
Net protein requirement for growth (g/d).
}
\description{
NASEM (2021) equation to predict the net protein requirement for growth - Eq. 10-14.
}
\examples{
net_protein_gain(EBW_gain = 0.4, retained_energy = 0.9)
}
