% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{NPgain}
\alias{NPgain}
\title{NASEM (2021) equation to predict the net protein requirement for growth - Eq. 10-14.}
\usage{
NPgain(EBWgain = 0.4, retained_energy = 0.9)
}
\arguments{
\item{EBWgain}{Empty body weight gain (kg/d).}

\item{retained_energy}{retained energy (Mcal/d).}
}
\value{
Net protein requirement for growth (g/d).
}
\description{
NASEM (2021) equation to predict the net protein requirement for growth - Eq. 10-14.
}
\examples{
NPgain(EBWgain = 0.4, retained_energy = 0.9)
}
