% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plots.R
\name{scenario_plot}
\alias{scenario_plot}
\title{Scenario plot.}
\usage{
scenario_plot(dataset, var_y)
}
\arguments{
\item{dataset}{A dataset with the variables to be plotted.}

\item{var_y}{A character string with the variable to be plotted.}
}
\value{
A plotly object.
}
\description{
Scenario plot.
}
