% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-protein-requirements.R
\name{scurf_cp}
\alias{scurf_cp}
\title{NASEM (2021) equation to predict the allowance for scurf - Eq. 10-10.}
\usage{
scurf_cp(BW = 40)
}
\arguments{
\item{BW}{Body weight (kg).}
}
\value{
The allowance for scurf (hair, skin, secretions, etc.; g/day).
}
\description{
NASEM (2021) equation to predict the allowance for scurf - Eq. 10-10.
}
\examples{
scurf_cp(BW = 45)
}
