% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-starter-intake.R
\name{starter_intake_nasem}
\alias{starter_intake_nasem}
\title{The NASEM (2021) equation to predict the average starter intake of calves.}
\usage{
starter_intake_nasem(BW = 45, MEiLD = 2.76, FPstarter = 2, temperature = 30)
}
\arguments{
\item{BW}{Body weight (kg).}

\item{MEiLD}{Metabolizable energy content of the liquid diet (Mcal/kgDM).}

\item{FPstarter}{Week relative the first week of starter offering (weeks).}

\item{temperature}{Average temperature (°C).}
}
\value{
The expected average starter intake of a calf (kg).
}
\description{
The NASEM (2021) equation to predict the average starter intake of calves.
}
\examples{
starter_intake_nasem(BW = 50, FPstarter = 5, temperature = 20, MEiLD = 2.76)
}
