% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-starter-intake.R
\name{starter_intake_silva2019}
\alias{starter_intake_silva2019}
\title{The Silva et al. (2019) equation to predict average starter intake of calves (kg).}
\usage{
starter_intake_silva2019(milk_intake = 5, age = 30)
}
\arguments{
\item{milk_intake}{Milk intake (l).}

\item{age}{Calf age (days).}
}
\value{
The expected average starter intake of a calf (kg).
}
\description{
The Silva et al. (2019) equation to predict average starter intake of calves (kg).
}
\examples{
starter_intake_silva2019(milk_intake = 5, age = 30)
}
\seealso{
The equation used here is also available in \href{https://www.sciencedirect.com/science/article/pii/S1751731118000666}{Silva et al. (2019)}.
}
