% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-calf-starter-me.R
\name{starter_met_energy}
\alias{starter_met_energy}
\title{Metabolizable energy of calf starter (Mcal/kg DM).}
\usage{
starter_met_energy(
  cs_ndf = 12,
  cs_nfc = 50,
  cs_cp = 22,
  cs_ee = 6,
  nfc_intake = 15,
  form_of_starter = "pelleted"
)
}
\arguments{
\item{cs_ndf}{Neutral detergent fiber of calf starter (\% DM).}

\item{cs_nfc}{Non-fiber carbohydrates of calf starter (\% DM).}

\item{cs_cp}{Crude protein of calf starter (\% DM).}

\item{cs_ee}{Ether extract of calf starter (\% DM).}

\item{nfc_intake}{Cumulative intake of non-fiber carbohydrates (\% DM).}

\item{form_of_starter}{Form of calf starter (pelleted or texturized).}
}
\value{
Metabolizable energy of calf starter (Mcal/kg DM).
}
\description{
Metabolizable energy of calf starter (Mcal/kg DM).
}
\examples{
starter_met_energy(cs_ndf = 12,
                  cs_nfc = 50,
                  cs_cp = 22,
                  cs_ee = 6,
                  nfc_intake = 15,
                  form_of_starter = "pelleted")
}
