% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-get-calf-requirements.R
\name{get_calf_requirements}
\alias{get_calf_requirements}
\title{Title}
\usage{
get_calf_requirements(
  liqDiet = rep(6, 70),
  liqDietME = 4.6,
  nfc_cs = 66,
  liqDietDM = 0.12,
  initBW = 45,
  weaningAge = 70,
  averTemp = 20,
  liqDietOnly = FALSE,
  mature_weight = 750,
  max_size = 100
)
}
\arguments{
\item{max_size}{}
}
\description{
Title
}
