% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-get-calf-requirements.R
\name{get_calf_requirements}
\alias{get_calf_requirements}
\title{Calculates the calf requirements.}
\usage{
get_calf_requirements(
  liqDiet = rep(6, 70),
  liqDietME = 4.6,
  starter_composition = list(cs_ndf = 12, cs_nfc = 50, cs_cp = 22, cs_ee = 6,
    form_of_starter = "pelleted"),
  cs_intake_equation = "NASEM",
  liqDietDM = 0.12,
  initBW = 45,
  weaningAge = 70,
  averTemp = 20,
  liqDietOnly = FALSE,
  mature_weight = 750,
  max_size = 100
)
}
\arguments{
\item{liqDiet}{A numeric vector with the liquid diet intake.}

\item{liqDietME}{A numeric vector with the liquid diet metabolizable energy.}

\item{starter_composition}{A list with the composition of the starter. The list
has to contain cs_ndf, cs_nfc, cs_cp, and cs_ee.}

\item{cs_intake_equation}{is the equation used to calculate the intake of the starter.}

\item{liqDietDM}{A numeric vector with the dry matter content of the liquid diet.}

\item{initBW}{A numeric vector with the initial body weight.}

\item{weaningAge}{A numeric vector with the weaning age.}

\item{averTemp}{A numeric vector with the average temperature.}

\item{liqDietOnly}{A logical vector indicating if the calf is fed only liquid diet.}

\item{mature_weight}{A numeric vector with the mature weight.}

\item{max_size}{A numeric vector with the maximum size of the vectors.}
}
\value{
A dataframe with the calf requirements.
}
\description{
Calculates the calf requirements.
}
\examples{
get_calf_requirements(liqDiet = rep(6, 70),
                     liqDietME = 4.6,
                     starter_composition = list(
                       cs_ndf = 12,
                       cs_nfc = 50,
                       cs_cp = 22,
                       cs_ee = 6,
                       form_of_starter = "pelleted"),
                     cs_intake_equation = "NASEM",
                     liqDietDM = 0.12,
                     initBW = 45,
                     weaningAge = 70,
                     averTemp = 20,
                     liqDietOnly = FALSE,
                     mature_weight = 750,
                     max_size = 100)
}
