% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-calf-starter-me.R
\name{starter_met_energy}
\alias{starter_met_energy}
\title{Metabolizable energy calcualtion based on .}
\usage{
starter_met_energy(
  cs_ndf = 12.9,
  cs_nfc = 55.79,
  cs_cp = 22,
  cs_ee = 3.9,
  CSNFCI = 15
)
}
\arguments{
\item{cs_ndf}{Calf starter NDF content (\%).}

\item{cs_nfc}{Calf starter NFC content (\%).}

\item{cs_cp}{Calf starter CP content (\%).}

\item{cs_ee}{Calf starter EE content (\%).}

\item{CSNFCI}{Calf starter cumulative NFC intake (kg).}
}
\value{
Metabolizable energy (ME) of calf starter (Mcal/kg DM).
}
\description{
Metabolizable energy calcualtion based on .
}
\examples{
starter_met_energy(cs_ndf = 12.9,
                  cs_nfc = 55.79,
                  cs_cp = 22,
                  cs_ee = 3.9,
                  CSNFCI = 15)
}
