% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-cards.R
\name{valueBoxCustom}
\alias{valueBoxCustom}
\title{Cards to be shown.}
\usage{
valueBoxCustom(
  card_title = "I got",
  card_value = "teste1",
  card_icon = "emoji-smile",
  card_color = "#00891A",
  card_bottom_message = "Tatata",
  message_icon = "emoji-smile"
)
}
\arguments{
\item{card_title}{Box title.}

\item{card_value}{Box value.}

\item{card_icon}{Box icon.}

\item{card_color}{Box color.}

\item{card_bottom_message}{Box bottom message.}

\item{message_icon}{Message icon to be shown.}
}
\value{
Value in the box.
}
\description{
Cards to be shown.
}
